new: Enttec::Gom::Session
this loads the DMX GOM node and parses device_file and values from it.
For the tests it uses a mocked version of the Gom::Client. Gom::Client
is my take on a Gom::Client which i develop in this package but will get
separated once it proves usable beyond the enttec daemon.